model, texture updates, enums namespace
The Model class is now complete with simple functionality with room in the
future for expansion. The Model can load vertices, texture coordinates,
and indices for vertices. The Painter does not use the Model class yet,
but it will in the future

The GraphicsContext now holds resources for Models similar to how it holds
resources for Texture

Texture's constructors are gone and replaced with static functions to
allow for more customization when creating Textures. The Texture Color
constructor still exists.

Texture::getSolidColor() returns a lazy-initialized cached texture
consisting of a single white pixel.

Texture::getGradient() returns a lazy initialized cached texture of height
100 where pixel (0,0) is equal to 0.0f and interpolated vertically, meaning pixel (0, 99) is equal to 1.0.

WindowModule and the Input namespace is now part of the gfx namespace as opposed to the os
namespace

Enums previously declared in classes are now located in the Enums
namespace, under the gfx namespace.