removed all public instances of bind(), and remove Bindable
The reasoning why is because bind() on a texture or model is an
implementation specific thing. For OpenGL, it makes sense, but for other
graphics renders (Vulkan or DirectX) it doesn't make much sense.

Plus, low level operations like that should be handled by the underlying
GraphicsContext, not by the end user. The end user should not be
dealing with low level graphics or memory management like that.