rotations demo can change the amount of stars, no more destroy()
Warning: very broken release!
There were constant issues regarding destroy(), like being called while
an object is still alive, or being called after an object was already
deleted. I've started work towards getting everything to be in a
destructor instead, which not only solves a lot of those problems, but
is also the proper way to do things in C++ (RAII). It's also much safer
and in general as memory will always be cleared